+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.8.0
-- `Moved EntityQueryLanguage.DataApi` namespace to `EntityQueryLanguage.GraphQL` to better communicate that its intent is to adds\ GraphQL support
-- Add support the GraphQL arguments. See updated readme
-- Move data results to field `data` in the resulting object to match GraphQL
+- `Moved EntityQueryLanguage.DataApi` namespace to `EntityQueryLanguage.GraphQL` to better communicate that its intent is to support
+- Add support GraphQL arguments in fields. See updated readme
+- By default `SchemaBuilder.FromObject<TType>()` generates a non-pural field for any type with a public `Id` property, with the argument name of `id`. E.g. A field `people` that returns a `IEnumerable<Person>` will result in a `person(id)` field
+- Move results to field `data` in the resulting object to match GraphQL
 - Support the `query` keyword for graphql
 
 # 0.7.0